requirements.txt
asgiref==3.5.1
Django==3.2.13
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.2.0
